Move BackupRangePartitionedProgress* files
