+++ fdbserver/backupworker/BackupWorkerRangePartitioned.cpp
@@ -25,7 +25,7 @@
 #include "fdbclient/SystemData.h"
 #include "fdbclient/Tracing.h"
 #include "BackupPartitionMap.h"
-#include "fdbserver/core/BackupRangePartitionedProgress.h"
+#include "fdbserver/backupworker/BackupRangePartitionedProgress.h"
 #include "fdbserver/core/Knobs.h"
 #include "fdbserver/core/LogSystem.h"
 #include "fdbserver/core/WaitFailure.h"
